requirements.txt
python-dotenv
openai
requests
google-genai
huggingface_hub
langchain
langchain-huggingface
langchain-google-genai
langchain-openai
langchain-aws
pydantic
mermaid-py
scikit-learn
langchain-chroma
langchain-community
langchain-pinecone
ipywidgets
langchain-postgres
psycopg[binary]


#for testing
pytest
pytest-mock